Разбил OfferPage на компоненты
